requirements.txt
asgiref==3.7.2
CacheControl==0.13.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
Django==4.2.5
djangorestframework==3.14.0
firebase-admin==6.2.0
google-api-core==2.11.1
google-api-python-client==2.99.0
google-auth==2.23.0
google-auth-httplib2==0.1.1
google-cloud-core==2.3.3
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.60.0
grpcio==1.58.0
grpcio-status==1.58.0
httplib2==0.22.0
idna==3.4
msgpack==1.0.5
proto-plus==1.22.3
protobuf==4.24.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.1.1
python-decouple==3.8
pytz==2023.3.post1
requests==2.31.0
rsa==4.9
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
